general fixes

+++ progress_bar.hpp
@@ -2,6 +2,6 @@
 #define GUARD_progress_bar_hpp
 
 // function to display the progress bar (input is a number from 0 to 1)
-void show_progress_bar(float progress);
+void show_progress_bar(double progress);
 
 #endif
